# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "text-classifier",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Task",
        "operationId": "task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDefinition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/texts/next": {
      "post": {
        "summary": "Next Text",
        "operationId": "next_text_api_texts_next_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextTextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/annotations": {
      "post": {
        "summary": "Annotations",
        "operationId": "annotations_api_annotations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/stats/me": {
      "get": {
        "summary": "Stats Me",
        "operationId": "stats_me_api_stats_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/stats/leaderboard/{task_id}": {
      "get": {
        "summary": "Stats Lb",
        "operationId": "stats_lb_api_stats_leaderboard__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnnotationSubmit": {
        "properties": {
          "text_id": {
            "type": "string",
            "title": "Text Id"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/TaskAnnotation"
            },
            "type": "array",
            "title": "Annotations"
          }
        },
        "type": "object",
        "required": [
          "text_id",
          "annotations"
        ],
        "title": "AnnotationSubmit"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NextTextRequest": {
        "properties": {
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 1,
            "title": "Task Ids"
          }
        },
        "type": "object",
        "required": [
          "task_ids"
        ],
        "title": "NextTextRequest"
      },
      "NextTextResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "language"
        ],
        "title": "NextTextResponse"
      },
      "TaskAnnotation": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "selected_classes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Classes"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "selected_classes",
          "start_time",
          "end_time"
        ],
        "title": "TaskAnnotation"
      },
      "TaskClass": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label_en": {
            "type": "string",
            "title": "Label En"
          },
          "label_cs": {
            "type": "string",
            "title": "Label Cs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label_en",
          "label_cs"
        ],
        "title": "TaskClass"
      },
      "TaskDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description_md": {
            "type": "string",
            "title": "Description Md"
          },
          "multi_choice": {
            "type": "boolean",
            "title": "Multi Choice"
          },
          "max_choices": {
            "type": "integer",
            "title": "Max Choices",
            "default": 1
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/TaskClass"
            },
            "type": "array",
            "title": "Classes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description_md",
          "multi_choice",
          "classes"
        ],
        "title": "TaskDefinition"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}